requirements.txt
google-api-python-client==2.65.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
google~=3.0.0
python-dotenv~=0.21.0